14장. 바꾸기 - 원할때 필요할 때 환율을 얻기위한 내용

+++ src/main/java/com/example/tdd/domain/Bank.java
@@ -11,8 +11,8 @@ public class Bank {
     }
 
     int rate(String from, String to){
-        int rate = (from.equals("CHF") && to.equals("USD")) ? 2 : 1;
-        return rate;
+        Integer rate = (from.equals("CHF") && to.equals("USD")) ? 2 : 1;
+        return rate.intValue();
     }
 
     public void addRate(String from, String to, int rate) {
